The message opener's style is a short
The permuter got BtlOpenMessage from 92.5% to 95.9% by narrowing the style
argument, though it did it by misdeclaring BtlBoxOpen in this translation unit
- box.c defines that parameter as an int. The honest version is the same
narrowing one step earlier: BtlOpenMessage's own style parameter is a short,
and the sign extension the original does before passing it on is what says so.

What is left is the frame, which the original makes eight bytes larger with
the same four saved registers, and the register assignment that goes with it.
Locals for the adjusted coordinates were the obvious candidate and changed
nothing. Re-seeded the permuter from the better base.

All 11 targets byte-exact; btlp unchanged at 156 functions.

+++ src/p1-jp/btlp/openmessage.c
@@ -45,7 +45,7 @@ extern u_char    g_btl_fast_anim;
 extern int  BtlTextOpen(const u_char *script, short x, short y);
 extern void BtlBoxOpen(short cols, short x, short y, int style);
 
-void BtlOpenMessage(int flags, int style, const u_char *script, short x,
+void BtlOpenMessage(int flags, short style, const u_char *script, short x,
                     short y)
 {
     int cols;
